+++ client/components/containers/card/ArtistCard.vue
@@ -53,7 +53,7 @@
 import Vue, { PropType } from 'vue';
 import { RootState } from 'typed-vuex';
 
-import UserAvatar, { MediaIcon } from '~/components/parts/avatar/UserAvatar.vue';
+import UserAvatar, { MediaIcon } from '~/components/parts/image/UserAvatar.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
 import { SpotifyAPI } from '~~/types';
