+++ client/components/containers/card/PlaylistCard.vue
@@ -56,7 +56,7 @@
 import Vue, { PropType } from 'vue';
 import { RootState } from 'typed-vuex';
 
-import ReleaseArtwork, { MediaIcon } from '~/components/parts/avatar/ReleaseArtwork.vue';
+import ReleaseArtwork, { MediaIcon } from '~/components/parts/image/ReleaseArtwork.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
 import { App, SpotifyAPI } from '~~/types';
