+++ client/components/parts/image/AvatarOverlay.vue
