+++ client/components/parts/image/ReleaseArtwork.vue
@@ -1,34 +1,29 @@
 <template>
   <v-hover #default="{ hover }">
-    <v-avatar
+    <v-img
       v-if="src != null"
+      :src="src"
+      :alt="alt"
       :min-width="minSize || size"
       :min-height="minSize || size"
-      :size="size"
+      :width="size"
+      :height="size"
       :max-width="maxSize || size"
       :max-height="maxSize || size"
+      :aspect-ratio="1"
       :class="{ 'g-box-shadow': shadow }"
+      @load="onLoaded"
     >
-      <v-img
-        :src="src"
-        alt="user-avaar"
-        :min-width="minSize || size"
-        :min-height="minSize || size"
-        :height="size"
-        :width="size"
-        :max-height="maxSize || size"
-        :max-width="maxSize || size"
-        :aspect-ratio="1"
-        @load="onLoaded"
-      >
+      <template v-if="isOverlayed">
         <AvatarOverlay
-          :hover="isOverlayed && hover"
+          v-show="hover"
+          :hover="hover"
           :size="size"
           :icon="icon"
           @on-clicked="onClicked"
         />
-      </v-img>
-    </v-avatar>
+      </template>
+    </v-img>
 
     <v-sheet
       v-else
@@ -38,11 +33,10 @@
       :height="size"
       :max-width="maxSize || size"
       :max-height="maxSize || size"
-      :class="$style.UserAvatar__noAvatar"
-      class="rounded-circle"
+      :class="$style.ReleaseArtwork__noArtwork"
     >
-      <v-icon :size="noAvatarIconSize">
-        {{ defaultUserIcon }}
+      <v-icon :size="noArtworkIconSize">
+        mdi-music
       </v-icon>
     </v-sheet>
   </v-hover>
@@ -50,33 +44,20 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue';
-import AvatarOverlay from '~/components/parts/avatar/AvatarOverlay.vue';
+import AvatarOverlay from '~/components/parts/image/AvatarOverlay.vue';
 
 export type MediaIcon = 'mdi-play-circle' | 'mdi-pause-circle'
 
-type Data = {
-  noAvatarIconSize: number
+export type Data = {
+  noArtworkIconSize: number
 }
 
 const ON_MEDIA_BUTTON_CLICKED = 'on-media-button-clicked';
 const ON_LOADED = 'on-loaded';
 
 export type On = {
-  ON_MEDIA_BUTTON_CLICKED: void
-  ON_LOADED: void
-}
-
-export type Props = {
-  src: string | undefined
-  alt: string
-  size?: number
-  minSize?: number
-  maxSize?: number
-  isOverlayed?: boolean
-  icon?: MediaIcon
-  defaultUserIcon?: string
-  smallIcon?: boolean
-  shadow?: boolean
+  [ON_MEDIA_BUTTON_CLICKED]: void
+  [ON_LOADED]: void
 }
 
 export default Vue.extend({
@@ -105,19 +86,11 @@ export default Vue.extend({
       type: Number as PropType<number | undefined>,
       default: undefined,
     },
-    isOverlayed: {
-      type: Boolean,
-      default: false,
-    },
     icon: {
       type: String as PropType<MediaIcon>,
       default: 'mdi-play-circle',
     },
-    defaultUserIcon: {
-      type: String,
-      default: 'mdi-account-circle-outline',
-    },
-    smallIcon: {
+    isOverlayed: {
       type: Boolean,
       default: false,
     },
@@ -129,17 +102,12 @@ export default Vue.extend({
 
   data(): Data {
     const baseSize = this.size ?? this.minSize;
-    let noAvatarIconSize: number;
-    if (baseSize == null) {
-      noAvatarIconSize = 60;
-    } else {
-      noAvatarIconSize = this.smallIcon
-        ? baseSize * 0.4
-        : baseSize;
-    }
+    const noArtworkIconSize = baseSize != null
+      ? baseSize * 0.4
+      : 60;
 
     return {
-      noAvatarIconSize,
+      noArtworkIconSize,
     };
   },
 
@@ -155,8 +123,8 @@ export default Vue.extend({
 </script>
 
 <style lang="scss" module>
-.UserAvatar {
-  &__noAvatar {
+.ReleaseArtwork {
+  &__noArtwork {
     display: flex;
     justify-content: center;
     align-items: center;
