+++ client/components/parts/list/ContentListItem.vue
@@ -55,8 +55,8 @@
 import Vue, { PropType } from 'vue';
 import { RawLocation } from 'vue-router';
 
-import ReleaseArtwork from '~/components/parts/avatar/ReleaseArtwork.vue';
-import UserAvatar from '~/components/parts/avatar/UserAvatar.vue';
+import ReleaseArtwork from '~/components/parts/image/ReleaseArtwork.vue';
+import UserAvatar from '~/components/parts/image/UserAvatar.vue';
 import ArtistNames from '~/components/parts/text/ArtistNames.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
 import { TRACK_LIST_ARTWORK_SIZE } from '~/constants';
