+++ client/components/parts/text/ArtistNames.vue
@@ -53,7 +53,7 @@
 <script lang="ts">
 import Vue, { PropType } from 'vue';
 
-import UserAvatar from '~/components/parts/avatar/UserAvatar.vue';
+import UserAvatar from '~/components/parts/image/UserAvatar.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
 import { App, SpotifyAPI } from '~~/types';
