+++ client/pages/artists/_artistId.vue
@@ -140,7 +140,7 @@ import { Component, Vue } from 'nuxt-property-decorator';
 import { Context } from '@nuxt/types';
 import { RootState, ExtendedMutationPayload } from 'typed-vuex';
 
-import UserAvatar from '~/components/parts/avatar/UserAvatar.vue';
+import UserAvatar from '~/components/parts/image/UserAvatar.vue';
 import HashTags from '~/components/parts/chip/HashTags.vue';
 import ContextMediaButton, { On as OnMediaButton } from '~/components/parts/button/ContextMediaButton.vue';
 import FollowButton, { On as OnFollow } from '~/components/parts/button/FollowButton.vue';
