+++ client/pages/episodes/_episodeId.vue
@@ -117,7 +117,7 @@
 import { Vue, Component } from 'nuxt-property-decorator';
 import { RootState } from 'typed-vuex';
 
-import ReleaseArtwork from '~/components/parts/avatar/ReleaseArtwork.vue';
+import ReleaseArtwork from '~/components/parts/image/ReleaseArtwork.vue';
 import ExplicitChip from '~/components/parts/chip/ExplicitChip.vue';
 import ContextMediaButton, { On as OnMediaButton } from '~/components/parts/button/ContextMediaButton.vue';
 import ReleaseDate from '~/components/parts/text/ReleaseDate.vue';
